refactor: migrate scripts/governance.py to src/quality/ (#344)
T2 of ADR-0010 implementation. Moves governance to src/quality/
and updates callers:
- agents/writer_agent.py
- agents/research_agent.py
- scripts/economist_agent.py (additional caller beyond brief, since this
  KEEP-script also bare-imports governance)

Refs #344
